Create receiveAttack function and tests
I added a receiveAttack function to the gameboardFactory that is
returned as part of the gameboard. It takes a pair of coordinates and
determines if it hits a ship or misses and records the coordinates in
the shots object. I also added a shipsLog array so that receiveAttack
can call hit on the ship.