Create placeShipsUI function
The placeShipsUI function allows a user to select which places on the
board they want their ships to be. It displays an error when the user
attempts to place a ship on a place that is already taken. Moving from
placing ships to the next phase of the game still needs work.