Create some intelligence for CPU hit logic
I updated the createShotCoordinates function that the CPU uses for
determining what coordinates to aim at. It will now look at what hits
have happened and try to guess a coordinate that is one away in the x
or y direction. It also ignores hits that are for sunken ships.